Bug fixes and tests
-Fixed bug for weather locations not showing up for precipitation locations
-added marquee text for additional text overflows
- added repository tests to simulate offline tests

+++ app/src/main/java/com/brian/weather/data/remote/dto/Forecast.kt
@@ -66,7 +66,7 @@ data class ForecastForDay(
 
 @Serializable
 data class Hour(
-    val time_epoch: Int,
+    val time_epoch: Long,
     val time: String,
     val temp_f: Double,
     val temp_c: Double,
